complete app!
- added manager pages, to add a personal list. (added a few extra functions to the managers list then what a normal user has, like going back to the managers menu)

- fixed bugs where the app would crush, the user crashed when trying to add a products to pending list.
the manager crashed on every other thing.

the solution: personal list uses the username to create the list so we need to transfer the username from page to page. if not, then when moving from a page that doesnt use the username to a page that does, the page would try to get the username but it doesnt not exist in the intent. so the app fails.

so, summery:
fixed bugs, added functions to manager.