100% DONE APP.
added voice to text option.
when adding an item to pending list/ regular list there is an option to talk insted of write.

added dialogs for user and manager. they open up after adding an item to regular or pending list.

any change from now on should potentialy be cosmetic.